src/requirements.txt
python-telegram-bot==20.8
requests==2.31.0
feedparser==6.0.10
apscheduler==3.10.3
